refactor page names
