+++ src/components/AppBar.tsx
@@ -12,7 +12,7 @@ import Tooltip from '@mui/material/Tooltip';
 import MenuItem from '@mui/material/MenuItem';
 import { Link } from 'react-router-dom';
 
-const pages = ['User', 'Workout'];
+const pages = ['User', 'Training'];
 const settings = ['Login', 'Register'];
 // const settings = ['Profile', 'Account', 'Dashboard', 'Login', 'Logout', 'Register'];
