+++ src/pages/Training/trainingDetail.page.tsx
